cloud_functions/record_progress/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore
